WIP: Wikidata Dev branch progress (to demo)
This is where we are at with the Wikidata dev branch so far. We'll
cut the demo from here and make a handful of surface changes to
that branch including quieting the output, and creating skeletons
etc. We'll continue to use the integration branch for dev purposes
so this branch only exists for demonstration.