Add additional Skeleton tests to Wikidata work
We provide some additional tests to make sure results come back
from the identifier as expected. With this addition we add checks
for extension mismatches and container pattern matching when the
Wikidata signature is combined with PRONOM.

The wikidata definitions file has been minimized in an attempt to
improve test execution time.