Remove source field from Wikidata results
Removing one of the two ways of displaying source (provenance)
information in the Wikidata result simplifies the Siegfried source
code and makes it easier to maintain.

The basis field could show provenance if explicitly selected by the
user. The source field would show the provenance of binary signatures
in a separate field -- inline, vs. separate field.

We elected to remove the source field, at least, for now, as it
was difficult to align the information in it with that of the `sf`
basis field. Where the basis field can hold other information without
provenance in its list and not necessarily at the same 'list'
indices.

The `-sourceinline` flag is removed as well as alternative branches
through the code for a separate source field.