Model siegfried output encoding current state
Siegfried is currently outputting incorrectly encoded data given
certain input data, e.g. control characters in file-paths. This
commit adds tests that confirm this. Tests pass for each scenario,
i.e. control character data when it fails validation, or valid data
when it passes. The commit can be used to verify the success of any
fix or refactor as it will start to fail when siegfried starts to
output entirely correct data. The test can be reworked to test for
True-Positives and not True-Negatives.