Enable users to configure a custom Wikibase
This commit sees the code added that is needed for users to specify
a custom Wikibase endpoint.

Given a custom endpoint, the caller will require a new SPARQL query
that can be specified in Siegfried %HOME% under wikibase.sparql.

A custom properties file is needed as well for properties to be
configured at runtime. Similarly, this can be configured in Siegfried
%HOME% in wikibase.json.

Significant testing has been put in place to ensure that both
Wikidata and Wikibase pathways are handled inuitively for the user.

Mock tests have been put in place to properly test a custom Wikibase
harvest from building signatures to verifying results returned via
Siegfried's identification API.

Wikibase functionality will be better documented in the Siegfried
docs.