Create in-memory filesystem for PRONOM skeletons
We can avoid writing to disk and make the tests here more portable by
reading from an in-memory filesystem. The skeletons themselves are
small and so can be easily stored in-line as strings and then turned
into byte objects.

Given the refactor to in-memory objects, we also take the opportunity
to add a file that won't identify with the minimal PRONOM signature
file and PRONOM reports. Type should be a nil-string as with many
of the other fields.