partially revert glob support to Windows-only fallback
The support for glob arguments introduced in v1.10.0 led to regressions reported in https://github.com/richardlehane/siegfried/issues/227.

Special characters used for glob patterns are also allowed in filenames on Linux and to a reduced extent in Windows, consequently such files couldn't be matched directly anymore as before without escaping.

This switches the treatment of arguments back to the pre-1.10 behavior of doing a direct match first. Glob support is retained as an attempted fallback on Windows when no direct match could be found.