Improve handling of droidXML config
A number of calls are made to retrieve the default location of the
DROID signature file dynamically. These are turned into a single
call in the app init. Further, there is a bit more feedback during
harvest to provide the user with more info when a DROID signature
file does not exist.