Introduce more context to logging
A new log formatter is introduced that adds program contex to the
log output, e.g. source file, and line of code where a log call
is made. Additional program context can be added consistently.
Datetime is formatted in ISO format and an option is provided to
use UTC or local timezones.